+++ app/src/main/java/com/shojishunsuke/kibunnsns/clean_arc/data/repository/LanguageAnalysisRepository.kt
@@ -1,8 +1,5 @@
 package com.shojishunsuke.kibunnsns.clean_arc.data.repository
 
 interface LanguageAnalysisRepository {
-
-   suspend fun getScore(text:String):Int
-
-
+   suspend fun getScore(text:String):Float
 }
