+++ app/src/main/java/com/shojishunsuke/kibunnsns/model/Post.kt
@@ -8,7 +8,7 @@ import java.util.*
 @IgnoreExtraProperties
 data class Post(
     val contentText:String = "",
-    var sentiScore :Int = 0,
+    var sentiScore :Float = 0f,
     val date : Date = Date(),
     val actID  : String = "soccer"
 ){}
